server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-chess==1.999
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0

